Add support for mqtt_subscribe
In order to reuse most of our code for other projects, we want to special
case salt/engines/hook where we need to parse the incoming payload. For
other types of events, we're going to assume that the data is already in
the correct format and does not need extra parsing.

This will allow other projects like https://github.com/kfdm/salt-mqtt to
deploy the correct payload directly into Salt's bus and not worry about
extra special casing there.